refactor(dashboard-auth): move board views to admin module
