+++ apps/dashboard/src/modules/admin/components/banners/BannerDialog.vue
