+++ apps/dashboard/src/modules/admin/components/banners/BannerItem.vue
@@ -43,13 +43,13 @@
       </div>
     </div>
   </div>
-  <BannerDialogComponent v-model:visible="isEditDialogVisible" v-model:banner="banner" />
+  <BannerDialog v-model:visible="isEditDialogVisible" v-model:banner="banner" />
 </template>
 <script setup lang="ts">
 import Image from 'primevue/image';
 import Tag from 'primevue/tag';
 
-import BannerDialogComponent from '@/components/BannerDialogComponent.vue';
+import BannerDialog from '@/modules/admin/components/banners/BannerDialog.vue';
 
 import type { BannerResponse } from '@sudosos/sudosos-client';
 import { computed, ref } from 'vue';
