+++ apps/dashboard/src/modules/admin/views/AdminBannersView.vue
@@ -1,11 +1,11 @@
 <template>
     <div class="page-container">
         <div class="page-title">{{ $t('banner.Banners') }}</div>
-        <BannersListComponent :banners="banners" :skip="0" :take="5"/>
+        <BannersList :banners="banners" :skip="0" :take="5"/>
     </div>
 </template>
 <script setup lang="ts">
-import BannersListComponent from '@/components/BannersListComponent.vue';
+import BannersList from '@/modules/admin/components/banners/BannersList.vue';
 import { onMounted } from 'vue';
 import { useBannersStore } from '@/stores/banner.store';
 import { storeToRefs } from 'pinia';
